feat(kit-demo): Island tab + AspectRatio / broken-avatar / swiper-in-modal
- Move Island demos to their own tab (new IslandSection)
- Add AspectRatio block (renders as <image> for the 16:9 case)
- Add broken-image avatars to show initials/icon fallback
- Host an interactive Swiper inside the Modal
- IslandGroup: pin bottom/top via inline style so Lynx actually floats it
